requirements.txt
Django==3.2
django-allauth==0.44.0
django-rest-auth==0.9.5
drf-yasg==1.20.0
djangorestframework~=3.12.4
psycopg2==2.8.6
requests~=2.25.1
gunicorn
whitenoise
